Task 374: fix api endpoints
