+++ client/pom.xml
@@ -30,7 +30,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.8</version>
                 <configuration>
-                    <mainClass>ui.Main</mainClass>
+                    <mainClass>ui.scopechange.Main</mainClass>
                 </configuration>
             </plugin>
             <plugin>
